+++ Changelog.md
@@ -18,6 +18,7 @@ v2.40.0 (not published yet)
 ====================
 
 - UI: Fix path handling when passed to StaxRip via command line ([#1375](/../../issues/1375), [#1377](/../../issues/1377), [#1378](/../../issues/1378))
+- UI: Fix blocking script error messages during job proceeding ([#1231](/../../issues/1231))
 - UI: Add question when adding an existing and active job with similar project file path
 - SvtAv1EncApp: Add long path support ([#1374](/../../issues/1374))
 - Update tools